requirements.txt
# requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
fastapi>=0.104.1
uvicorn>=0.24.0
pynacl>=1.5.0
aiohttp>=3.8.1
